backend/requirements.dev.txt
autopep8==2.3.2
fastapi-cli[standard-no-fastapi-cloud-cli]==0.0.24
httpx==0.28.1
pycodestyle==2.14.0
pylint==4.0.5
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.1.0
